Move all cm-related code under a new 'cm' dir
