+++ src/cm/commands/getWorkspaceFromPath/commandInfo.ts
