+++ src/cm/shell/interfaces/cmResult.ts
