+++ src/cm/shell/lineStream.ts
