+++ src/test/suite/commands/checkin/checkinParser.test.ts
@@ -1,7 +1,7 @@
 import { expect } from "chai";
 import * as os from "os";
-import { ICmParser } from "../../../../cmShell";
-import { CheckinParser } from "../../../../commands/checkin/checkinParser";
+import { CheckinParser } from "../../../../cm/commands/checkin/checkinParser";
+import { ICmParser } from "../../../../cm/shell";
 import { ICheckinChangeset } from "../../../../models";
 
 describe("Checkin Parser", () => {
